requirements.txt
cbpro==1.1.4
python-dotenv==0.19.0
colorama==0.4.4
diskcache==5.2.1